(AKA-5) Переместил файлы в supporting files. Добавил контроллеры в табБарСонтроллер
