--- Map/AppDelegate.swift
@@ -1,24 +1,0 @@
-//
-//  AppDelegate.swift
-//  Map
-//
-//  Created by Arkadiy Grigoryanc on 16.12.2019.
-//
-
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-	var window: UIWindow?
-
-	func application(_ application: UIApplication,
-					didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-
-		window = UIWindow(frame: UIScreen.main.bounds)
-		window?.rootViewController = UITabBarController()
-		window?.makeKeyAndVisible()
-
-		return true
-	}
-}
