+++ Map/Supporting files/Assets.xcassets/AppIcon.appiconset/Contents.json
