+++ Map/Supporting files/Assets.xcassets/Contents.json
